Remove applePlugin from swiftui demo
