+++ MULTIPLATFORM.md
@@ -192,19 +192,9 @@ To use a locally built compose in [KMP with Compose wizard project](https://kmp.
 
 Now the project will build with the locally published Compose.
 
-### Run mpp/demo-uikit sample on iOS
+### Run mpp/demo-swiftui sample on iOS with Xcode
 
-- Install plugin **Kotlin Multiplatform Mobile** for AppCode
-- Disable Android Plugin in IDE preferences
-- Open root of repository with AppCode
-- Increase AppCode IDE memory more that 10GB
-- Restart AppCode
-- Wait while project synchronization with Gradle
-- [Optional] To run on real iOS device
-  - Device should be at least with iOS 16.5. And Xcode should be at least 14.3
-  - Find your TEAM_ID with instruction <https://github.com/JetBrains/compose-multiplatform-template#running-on-a-real-ios-device>
-  - Create file project.properties in the root of repository. Add property `TEAM_ID=[your team id]` without double quotes, for example `TEAM_ID=ABC123ABC1`.
-- Choose run configuration **iOS App**
+Open the iosApp.xcodeproj with XCode and press the Run button.
 
 ### Run mpp/demo sample on iOS with Xcode
