+++ compose/mpp/demo-swiftui/.gitignore
@@ -1,0 +1,2 @@
+iosApp/iosApp.xcodeproj/project.xcworkspace/xcuserdata/**
+iosApp/iosApp.xcodeproj/xcuserdata/**
