+++ compose/mpp/demo-swiftui/iosApp/Configuration/Config.xcconfig
@@ -1,0 +1,7 @@
+TEAM_ID=
+
+PRODUCT_NAME=KotlinProject
+PRODUCT_BUNDLE_IDENTIFIER=org.example.project.KotlinProject$(TEAM_ID)
+
+CURRENT_PROJECT_VERSION=1
+MARKETING_VERSION=1.0
